Changed endpoint (clean code):
1) /auth/userBoards -> /boards/get
2) /userboards/add -> /boards/add
