+++ backend/src/auth/auth.controller.ts
@@ -1,5 +1,4 @@
 import { Controller, Get, Post, Request, UseGuards } from '@nestjs/common';
-import { UserboardsService } from 'src/userboards/userboards.service';
 import { AuthService } from './auth.service';
 import { JwtAuthGuard } from './jwt-auth.guard';
 import { LocalAuthGuard } from './local-auth.guard';
@@ -7,8 +6,7 @@ import { LocalAuthGuard } from './local-auth.guard';
 @Controller('auth')
 export class AuthController {
     constructor(
-        private readonly authService: AuthService,
-        private readonly userBoardsService: UserboardsService,
+        private readonly authService: AuthService
     ) {}
 
     @Post('register')
@@ -28,11 +26,5 @@ export class AuthController {
     @Get('profile')
     getProfile(@Request() req) { // get data about logged user
         return req.user;
-    }
-
-    @UseGuards(JwtAuthGuard)
-    @Get('userBoards')
-    getUserBoards(@Request() req) { // get boards of logged user
-        return this.userBoardsService.getUserBoards(req.user.id);
     }
 }
