+++ backend/src/userboards/userboards.module.ts
@@ -1,9 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Board } from 'src/boards/boards.entity';
-import { BoardsModule } from 'src/boards/boards.module';
 import { User } from 'src/users/users.entity';
-import { UserboardsController } from './userboards.controller';
 import { Userboard } from './userboards.entity';
 import { UserboardsService } from './userboards.service';
 
@@ -12,9 +10,7 @@ import { UserboardsService } from './userboards.service';
     TypeOrmModule.forFeature([Userboard]),
     TypeOrmModule.forFeature([User]),
     TypeOrmModule.forFeature([Board]),
-    BoardsModule
   ],
-  controllers: [UserboardsController],
   providers: [UserboardsService],
   exports: [UserboardsService],
 })
